Runtime Mapping Rotation

+++ code/game/machinery/hologram.dm
@@ -253,7 +253,7 @@ For the other part of the code, check silicon say.dm. Particularly robot talk.*/
 	hologram.anchored = 1//So space wind cannot drag it.
 	if(caller_id)
 		hologram.SetName("[caller_id.name] (Hologram)")
-		hologram.loc = get_step(src,1)
+		hologram.forceMove(get_step(src,1))
 		masters[caller_id] = hologram
 	else
 		hologram.SetName("[A.name] (Hologram)") //If someone decides to right click.
@@ -348,6 +348,8 @@ For the other part of the code, check silicon say.dm. Particularly robot talk.*/
 
 //Destruction procs.
 /obj/machinery/hologram/ex_act(severity)
+	if(atom_flags & ATOM_FLAG_INDESTRUCTIBLE)
+		return
 	switch(severity)
 		if(1.0)
 			qdel(src)
